requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
python-dotenv>=1.0.0
boto3>=1.28.0
pydantic>=2.0.0
streamlit>=1.32.0
